modify usage of keypool

+++ src/wallet/wallet.cpp
@@ -3320,7 +3320,7 @@ bool CWallet::TopUpKeyPool(unsigned int kpSize)
         if (kpSize > 0)
             nTargetSize = kpSize;
         else
-            nTargetSize = std::max(int64_t(gArgs.GetArg("-keypool", DEFAULT_KEYPOOL_SIZE)), int64_t(0));
+            nTargetSize = std::max(gArgs.GetArg("-keypool", DEFAULT_KEYPOOL_SIZE), (unsigned int)(0));
 
         // count amount of available keys (internal, external)
         // make sure the keypool of external and internal keys fits the user selected target (-keypool)
